foodback/requirements/prod.txt
# Everything needed in production

# Flask
click>=7.0
Flask==3.1.1
Werkzeug==3.1.3

# Database
Flask-SQLAlchemy==3.1.1
psycopg2-binary==2.9.10
PyMySQL==1.1.1
SQLAlchemy==2.0.42

# Migrations
Flask-Migrate==4.1.0

# Forms
email-validator==2.2.0
Flask-WTF==1.2.2
WTForms==3.2.1

# Deployment
gevent==25.5.1
gunicorn>=19.9.0
supervisor==4.2.5

# Flask Static Digest
Flask-Static-Digest==0.4.1

# Auth
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3

# Caching
Flask-Caching>=2.0.2

# Debug toolbar
Flask-DebugToolbar==0.16.0

# Environment variable parsing
environs==14.2.0

# JWT Token support
PyJWT==2.10.1

# CORS support
Flask-CORS==5.0.0

# File upload and validation
Pillow==11.1.0
python-magic==0.4.27; sys_platform != "win32"
python-magic-bin==0.4.14; sys_platform == "win32"

# Date/time utilities
python-dateutil==2.9.0

# Task scheduling
APScheduler==3.10.4

# AI and ML
google-generativeai==0.8.3
python-dotenv==1.0.1
